refactor: 로또 넘버 부생성자 제거

+++ src/main/kotlin/lotto/view/InputView.kt
@@ -15,13 +15,13 @@ object InputView {
         println("지난 주 당첨 번호를 입력해 주세요.")
         val input = readln()
         input.split(",").forEach { ValidationUtils.validateLottoNumber(it.trim()) }
-        return input.split(",").map { LottoNumber(it.trim()) }.toSet()
+        return input.split(",").map { LottoNumber(it.trim().toInt()) }.toSet()
     }
 
     fun getBonusNumber(): LottoNumber {
         println("보너스 번호를 입력해 주세요.")
         val input = readln()
         ValidationUtils.validateLottoNumber(input)
-        return LottoNumber(input)
+        return LottoNumber(input.toInt())
     }
 }
